texture coordinates go from 0 - 1 now.

+++ Template/renderContext.hpp
@@ -160,9 +160,9 @@ namespace graphics {
 
 			for (int i = x_min; i < x_max; ++i) {
 				float z = 1 / one_over_z;
-				ivec2 source = {
-					int(coord_over_z.x * z + 0.5), 
-					int(coord_over_z.y * z + 0.5)
+				vec2 source = {
+					(coord_over_z.x * z + 0.5), 
+					(coord_over_z.y * z + 0.5)
 				};
